remove temp classes and cleanup
change deprecated method get child position
