+++ sample/src/main/java/io/github/customadapterrecyclerview/MainActivity.kt
@@ -1,6 +1,5 @@
 package io.github.customadapterrecyclerview
 
-import android.annotation.SuppressLint
 import android.content.Intent
 import android.os.Bundle
 import android.widget.Button
